fix : Edit페이지에 있는 내용을 Post페이지로 옮기기
