+++ src/pages/Edit.jsx
@@ -10,86 +10,22 @@ import MessageCard from '@/components/MessageCard/MessageCard';
 import { getRecipientsId } from '@/api/recipients';
 import { BACKGROUND_COLOR_PALETTE } from '@/util/backgroundColors.jsx';
 import { deleteMessage } from '@/api/message';
+import Post from './Post';
 
 function Edit() {
-  const [isEdit, setIsEdit] = useState(false);
-  const handleEditClick = () => {
-    if (isEdit) {
-      console.log('true상태!!');
-    } else {
-      console.log('false상태!!');
-    }
-    setIsEdit(!isEdit);
-  };
-
-  const [postName, setPostName] = useState('');
-  const [postMessageCount, setPostMessageCount] = useState(0);
-  const [reactions, setReactions] = useState([]);
-  const [profileImages, setProfileImages] = useState([]);
-
-  const [messageContents, setMessageContents] = useState();
-  const [background, setBackground] = useState('var(--orange-200, #ffe2ad)');
-
-  const { recipientId } = useParams();
-
-  const handleRollingPaper = useCallback(async () => {
-    const results = await getRecipientsId({ id: recipientId });
-    const { name, messageCount, backgroundColor, backgroundImageURL, topReactions } = { ...results };
-    const { color } = BACKGROUND_COLOR_PALETTE[backgroundColor];
-    const recentMessages = [...results.recentMessages];
-
-    handlePostHeader(name, messageCount, topReactions, recentMessages);
-
-    setMessageContents(recentMessages);
-    setBackground({ color, backgroundImageURL });
-  }, [recipientId]);
-
-  useEffect(() => {
-    handleRollingPaper();
-  }, [handleRollingPaper]);
-
-  const handlePostHeader = (name, messageCount, topReactions, recentMessages) => {
-    setPostName(name);
-    setPostMessageCount(messageCount);
-    setReactions([...topReactions].slice(0, 3));
-    const recentPostProfileImages =
-      recentMessages.length === 0 ? [] : recentMessages.map((message) => message.profileImageURL).slice(0, 3);
-    setProfileImages(recentPostProfileImages);
-  };
-
-  const onDelete = useCallback(async (messageId) => {
-    await deleteMessage({ messageId });
-    setMessageContents((prevMessages) => prevMessages.filter((message) => message.id !== messageId));
-  });
+  // const [isEdit, setIsEdit] = useState(false);
+  // const handleEditClick = () => {
+  //   if (isEdit) {
+  //     console.log('true상태!!');
+  //   } else {
+  //     console.log('false상태!!');
+  //   }
+  //   setIsEdit(!isEdit);
+  // };
 
   return (
     <>
-      <PostHeader name={postName} messageCount={postMessageCount} reactions={reactions} profileImages={profileImages} />
-      <PostContainer $backgroundColor={background.color} $imageUrl={background.backgroundImageURL}>
-        {!isEdit && (
-          <EditButton $size="H40" type="button" onClick={handleEditClick}>
-            편집하기
-          </EditButton>
-        )}
-
-        <CardContainer>
-          {!isEdit && <PlusMessageCard />}
-          {messageContents &&
-            messageContents.map((messageCard) => (
-              <MessageCard
-                value={messageCard}
-                key={messageCard.id}
-                isEdit={isEdit}
-                onDelete={() => onDelete(messageCard.id)}
-              />
-            ))}
-        </CardContainer>
-        {isEdit && (
-          <SaveButtonContainer>
-            <SaveButton $size="big">저장하기</SaveButton>
-          </SaveButtonContainer>
-        )}
-      </PostContainer>
+      <Post />
     </>
   );
 }
@@ -164,36 +100,4 @@ const PostContainer = styled.div`
     grid-template-columns: repeat(3, 38.4rem);
     grid-template-rows: repeat(auto-fit, 28rem);
   }
-`;
-
-const SaveButtonContainer = styled.div`
-  bottom: 2.4rem;
-  position: absolute;
-  margin-top: 4.2rem;
-  padding: 2.4rem 2rem;
-  background-color: @media (min-width: 768px) {
-    margin-top: 13.2rem;
-    padding: 2.4rem;
-  }
-
-  @media (min-width: 1248px) {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 4rem;
-  }
-`;
-
-const SaveButton = styled(PrimaryButton)`
-  width: 100%;
-
-  @media (min-width: 1248px) {
-    width: 28rem;
-  }
-`;
-
-const EditButton = styled(OutlineButton)`
-  display: flex;
-  border-radius: 8px;
-  font-size: 1.6rem;
 `;
