+++ src/pages/Post.jsx
@@ -1,13 +1,29 @@
 import { useCallback, useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
+import { useParams, useLocation } from 'react-router-dom';
 import styled from 'styled-components';
 import PostHeader from '@/components/Header/PostHeader.jsx';
 import PlusMessageCard from '@/components/MessageCard/PlusMessageCard';
 import MessageCard from '@/components/MessageCard/MessageCard';
 import { getRecipientsId } from '@/api/recipients';
 import { BACKGROUND_COLOR_PALETTE } from '@/util/backgroundColors.jsx';
+import { deleteMessage } from '@/api/message';
+import PrimaryButton from '@/styles/button/PrimaryButton.jsx';
+import OutlineButton from '@/styles/button/OutlineButton.jsx';
 
 function Post() {
+  const location = useLocation();
+  const EditPage = location.pathname.includes('/edit');
+
+  const [isEdit, setIsEdit] = useState(false);
+  const handleEditClick = () => {
+    if (isEdit) {
+      console.log('true상태!!');
+    } else {
+      console.log('false상태!!');
+    }
+    setIsEdit(!isEdit);
+  };
+
   const [postName, setPostName] = useState('');
   const [postMessageCount, setPostMessageCount] = useState(0);
   const [reactions, setReactions] = useState([]);
@@ -43,14 +59,36 @@ function Post() {
     setProfileImages(recentPostProfileImages);
   };
 
+  const onDelete = useCallback(async (messageId) => {
+    await deleteMessage({ messageId });
+    setMessageContents((prevMessages) => prevMessages.filter((message) => message.id !== messageId));
+  });
+
   return (
     <>
       <PostHeader name={postName} messageCount={postMessageCount} reactions={reactions} profileImages={profileImages} />
+      {EditPage && !isEdit && (
+        <EditButton $size="H40" type="button" onClick={handleEditClick}>
+          편집하기
+        </EditButton>
+      )}
       <PostContainer $backgroundColor={background.color} $imageUrl={background.backgroundImageURL}>
-        <PlusMessageCard />
+        {!isEdit && <PlusMessageCard />}
         {messageContents &&
-          messageContents.map((messageCard) => <MessageCard value={messageCard} key={messageCard.id} />)}
+          messageContents.map((messageCard) => (
+            <MessageCard
+              value={messageCard}
+              key={messageCard.id}
+              isEdit={isEdit}
+              onDelete={() => onDelete(messageCard.id)}
+            />
+          ))}
       </PostContainer>
+      {isEdit && (
+        <SaveButtonContainer>
+          <SaveButton $size="big">삭제하기</SaveButton>
+        </SaveButtonContainer>
+      )}
     </>
   );
 }
@@ -95,4 +133,35 @@ const PostContainer = styled.div`
     grid-template-columns: repeat(3, 38.4rem);
     grid-template-rows: repeat(auto-fit, 28rem);
   }
+`;
+const SaveButtonContainer = styled.div`
+  bottom: 2.4rem;
+  position: absolute;
+  margin-top: 4.2rem;
+  padding: 2.4rem 2rem;
+  background-color: @media (min-width: 768px) {
+    margin-top: 13.2rem;
+    padding: 2.4rem;
+  }
+
+  @media (min-width: 1248px) {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    margin-top: 4rem;
+  }
+`;
+
+const SaveButton = styled(PrimaryButton)`
+  width: 100%;
+
+  @media (min-width: 1248px) {
+    width: 28rem;
+  }
+`;
+
+const EditButton = styled(OutlineButton)`
+  display: flex;
+  border-radius: 8px;
+  font-size: 1.6rem;
 `;
